Documents/Christian/CDD-main/requirements.txt
streamlit>=1.48.0
openai>=1.58.1
PyPDF2>=3.0.1
camelot-py[cv]>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
sap-ai-sdk-gen[all]>=5.10.0
langchain-core>=0.3.0
langchain-aws>=0.2.0

